Fixed sonar issues and code duplication
