+++ ksml-data-avro-apicurio/src/main/java/io/axual/ksml/data/notation/avro/apicurio/ApicurioAvroNotation.java
@@ -22,22 +22,16 @@ package io.axual.ksml.data.notation.avro.apicurio;
 
 import io.apicurio.registry.resolver.client.RegistryClientFacade;
 import io.axual.ksml.client.resolving.Resolver;
-import io.axual.ksml.data.exception.SchemaException;
-import io.axual.ksml.data.notation.avro.AvroNotation;
+import io.axual.ksml.data.notation.avro.RemoteSchemaAvroNotation;
 import io.axual.ksml.data.notation.vendor.VendorNotationContext;
-import io.axual.ksml.data.schema.DataSchema;
-import lombok.extern.slf4j.Slf4j;
 
 /**
  * Apicurio-backed AvroNotation that supports fetching schemas from a schema registry.
  * <p>
- * This extends AvroNotation to override {@link #fetchRemoteSchema(String, boolean)}, enabling
- * deferred schema resolution for stream definitions that omit an explicit schema name.
+ * Extends {@link RemoteSchemaAvroNotation}, providing the Apicurio-specific registry client and fetch.
  */
-@Slf4j
-public class ApicurioAvroNotation extends AvroNotation {
+public class ApicurioAvroNotation extends RemoteSchemaAvroNotation {
     private final RegistryClientFacade registryClient;
-    private final Resolver topicResolver;
 
     /**
      * Construct an AvroNotation with the provided vendor context.
@@ -45,32 +39,22 @@ public class ApicurioAvroNotation extends AvroNotation {
      * @param context the vendor notation context providing serde supplier, native mapper, and configs
      */
     public ApicurioAvroNotation(VendorNotationContext context, RegistryClientFacade registryClient, Resolver topicResolver) {
-        super(context);
+        super(context, topicResolver::resolve);
         this.registryClient = registryClient;
-        this.topicResolver = topicResolver;
     }
 
     @Override
-    public boolean supportsRemoteSchema() {
-        if (registryClient != null)
-            return true;
-        log.warn("Apicurio registry not configured, remote schema resolution is disabled");
-        return false;
+    protected boolean hasRegistryClient() {
+        return registryClient != null;
     }
 
     @Override
-    public DataSchema fetchRemoteSchema(String topic, boolean isKey) {
-        if (registryClient == null) {
-            throw new SchemaException("Cannot fetch remote schema: no schema registry client configured");
-        }
+    protected String registryDescription() {
+        return "Apicurio registry";
+    }
 
-        final var subject = topicResolver.resolve(topic) + (isKey ? "-key" : "-value");
-        try {
-            log.info("Fetching latest schema for subject '{}' from schema registry", subject);
-            final var schema = registryClient.getSchemaByGAV(null, subject, null);
-            return schemaParser().parse(subject, subject, schema);
-        } catch (Exception e) {
-            throw new SchemaException("Failed to fetch schema for subject '" + subject + "' from schema registry", e);
-        }
+    @Override
+    protected String fetchSchemaString(String subject) {
+        return registryClient.getSchemaByGAV(null, subject, null);
     }
 }
