+++ ksml-test-runner/src/main/java/io/axual/ksml/testrunner/TestDefinitionSchemaGenerator.java
@@ -21,6 +21,7 @@ package io.axual.ksml.testrunner;
  */
 
 import tools.jackson.core.JacksonException;
+import tools.jackson.databind.SerializationFeature;
 import tools.jackson.databind.json.JsonMapper;
 import tools.jackson.databind.node.ObjectNode;
 
@@ -47,7 +48,11 @@ import java.util.Map;
  */
 public class TestDefinitionSchemaGenerator {
 
-    private static final JsonMapper MAPPER = JsonMapper.builder().build();
+    // INDENT_OUTPUT keeps the generated docs/ksml-test-spec.json human-readable and consistent with the
+    // other generated spec files (Jackson 3 no longer pretty-prints by default).
+    private static final JsonMapper MAPPER = JsonMapper.builder()
+            .enable(SerializationFeature.INDENT_OUTPUT)
+            .build();
     private static final String DESCRIPTION = "description";
     private static final String TYPE_OBJECT = "object";
     private static final String REQUIRED = "required";
